refactor: EventSearchForm 리펙토링
